build: Add Nix Flake & derivation scaffolding
Also added a CI workflow for the Flake and derivation checks. Will be
upstreaming the derivation to Nixpkgs, but only on Deskflow releases.

Having the Flake & derivation in this repo allows people to build from
Git, or enter a development environment.

You will see `/.envrc` - this is for `direnv`, and if people have Nix
installed, it will automatically generate a developer environment,
Nixified.

One big benefit is that people can simply run the command:

`nix run github:deskflow/deskflow` - and Nix will fetch the latest
commit, build, and run Deskflow in one hit.

Later on, we should add 'Cachix' support, which means binaries will be
built and hosted on a Nix cache. This is free of charge.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>