build: Add Nix Flake & derivation scaffolding
Also added a CI workflow for the Flake and derivation checks.

Having the Flake & derivation in this repo allows people to build from
Git, or enter a development environment.

You will see `/.envrc` - this is for `direnv`, and if people have Nix
installed, it will automatically generate a developer environment,
Nixified.

I have also added `devenv` as the handler for the Nix
`devShells.default` output - this also builds a `.`devcontainer.json`,
which is completely self-contained, Nix-based, and contains all
libraries used by the Deskflow Nix derivation.

One big benefit is that people can simply run the command:

`nix run github:deskflow/deskflow` - and Nix will fetch the latest
commit, build, and run Deskflow in one hit.

Later on, we should add 'Cachix' support, which means binaries will be
built and hosted on a Nix cache. This is free of charge.

This is *not* merely a packaging effort. It also provides a developer
environment, as well as a Devcontainer.

Something of note, is that we're simply overriding the Deskflow
derivation from upstream Nixpkgs, which is overridden to use the
Deskflow Git repo from master as the source code.